backend/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-agents>=0.11.0",
    "python-dotenv>=1.2.2",
]
